% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bc_admin_regions}
\alias{bc_admin_regions}
\title{Get geometry object of the natural resource regions of British Columbia}
\usage{
bc_admin_regions()
}
\value{
A geometry object of the British Columbia natural resource zones
}
\description{
Get geometry object of the natural resource regions of British Columbia
}
