% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bc_bec}
\alias{bc_bec}
\title{Get geometry object of the biogeoclimatic zones of British Columbia}
\usage{
bc_bec()
}
\value{
A geometry object of the British Columbia biogeoclimatic zones
}
\description{
Get geometry object of the biogeoclimatic zones of British Columbia
}
