% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bc_bounds}
\alias{bc_bounds}
\title{Get geometry object of the provincial boundaries of British Columbia}
\usage{
bc_bounds(res = "low")
}
\arguments{
\item{res}{Whether the spatial resolution should be 'high' or 'low'}
}
\value{
A geometry object of the British Columbia provincial boundaries
}
\description{
Get geometry object of the provincial boundaries of British Columbia
}
