% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{get_collection_items}
\alias{get_collection_items}
\title{Get all items within a specific collection that the user has access to}
\usage{
get_collection_items(collection_id, token)
}
\arguments{
\item{collection_id}{The unique identifier of a collection discovered in \code{get_collections()}}

\item{token}{A session token generated by \code{get_token()} that is used to authenticate the request}
}
\value{
A dataframe containing information about the items with a specific collection
}
\description{
Get all items within a specific collection that the user has access to
}
