% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{get_collections}
\alias{get_collections}
\title{Get all collections that the user has access to}
\usage{
get_collections(token = NULL)
}
\arguments{
\item{token}{A session token generated by \code{get_token()} that is used to authenticate the request}
}
\value{
A dataframe containing information about the collections that the user has access to
}
\description{
Get all collections that the user has access to
}
