% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{get_data}
\alias{get_data}
\title{Get all items within a specific collection that the user has access to}
\usage{
get_data(item_id, token)
}
\arguments{
\item{token}{A session token generated by \code{get_token()} that is used to authenticate the request}

\item{card_id}{The unique identifier of an item discovered in \code{get_collection_items()}}
}
\value{
A dataframe containing information about the items with a specific collection
}
\description{
Get all items within a specific collection that the user has access to
}
