% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spi_data.R
\name{get_surveys}
\alias{get_surveys}
\title{Get survey summaries from the Species Inventory database through the BC Data Catalogue}
\usage{
get_surveys(species = NULL, region = NULL, crs = 3005)
}
\arguments{
\item{region}{The natural resource region to filter the results by}

\item{crs}{The coordinate reference system that the results should be returned in. Defauls to EPSG:3005 BC Albers.}
}
\description{
A wrapper around \code{bcdata::bcdata_get_data()} to retrieve all non-secure survey summaries from the Species Inventory database through the BC Data Catalogue

#' @param species The English name of a species to filter the results by
}
