% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spi_data.R
\name{get_telemetry}
\alias{get_telemetry}
\title{Get telemetry records from the Species Inventory database through the BC Data Catalogue}
\usage{
get_telemetry(species = NULL, start_date = NULL, end_date = NULL, crs = 3005)
}
\arguments{
\item{species}{The English name of a species to filter the results by}

\item{end_date}{The maximum date to filter the results by. Returns records before this date.}

\item{crs}{The coordinate reference system that the results should be returned in. Defauls to EPSG:3005 BC Albers.}

\item{start_end}{The minimum date to filter the results by. Returns records on or after this date.}
}
\description{
A wrapper around \code{bcdata::bcdata_get_data()} to retrieve all non-secure telemetry records from the Species Inventory database through the BC Data Catalogue
}
