% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{get_token}
\alias{get_token}
\title{Generate a session token to access the Metabase API for the BC Telemetry Warehouse}
\usage{
get_token(username, password)
}
\arguments{
\item{username}{Username or email or Metabase account}

\item{password}{Password of Metabase account}
}
\value{
A string to be passed as an argument to functions requiring it
}
\description{
Generate a session token to access the Metabase API for the BC Telemetry Warehouse
}
